// types.ts
export type RootStackParamList = {
	'(tabs)': undefined;
	journalContents: { title: string; content: string; tags: string[] };
	createJournalEntry: undefined;
	editProfile: undefined;
	buddies: undefined;
	otherProfile: undefined;
	'(auth)/welcome': undefined;
	'(auth)/signup': undefined;
	'(auth)/login': undefined;
	'+not-found': undefined;
};